separate form from dialog
